refactor: extract config module
